+++ cpp/example_code/dynamodb/CMakeLists.txt
@@ -4,8 +4,11 @@
 # Set the minimum required version of CMake for this project.
 cmake_minimum_required(VERSION 3.8)
 
+set(SERVICE_NAME dynamodb)
+set(SERVICE_COMPONENTS dynamodb)
+
 # Set this project's name.
-project("dynamodb-examples")
+project("${SERVICE_NAME}-examples")
 
 #Set the location of where Windows can find the installed libraries of the SDK.
 if(MSVC)
@@ -17,7 +20,9 @@ endif()
 set(CMAKE_CXX_STANDARD 11)
 
 # Enable CTest for testing these code examples.
-include(CTest)
+if(BUILD_TESTS)
+    include(CTest)
+endif()
 
 # Build shared libraries by default.
 if(NOT BUILD_SHARED_LIBS)
@@ -25,97 +30,76 @@ if(NOT BUILD_SHARED_LIBS)
 endif()
 
 # Find the AWS SDK for C++ package.
-find_package(AWSSDK REQUIRED COMPONENTS dynamodb)
+find_package(AWSSDK REQUIRED COMPONENTS ${SERVICE_COMPONENTS})
 
-# If the compiler is some version of Microsoft Visual C++, or another compiler simulating C++, 
+find_package(ZLIB)
+set(ADDITIONAL_LIBS )
+set(ADDITIONAL_INCLUDE_DIR)
+set(ADDITIONAL_DEFINES)
+
+if (ZLIB_FOUND)
+    set(ADDITIONAL_LIBS ${ZLIB_LIBRARIES})
+    set(ADDITIONAL_INCLUDE_DIR ${ZLIB_INCLUDE_DIRS})
+    set(ADDITIONAL_DEFINES _HAS_ZLIB_)
+endif()
+# If the compiler is some version of Microsoft Visual C++, or another compiler simulating C++,
 # and building as shared libraries, then dynamically link to those shared libraries.
 if(MSVC AND BUILD_SHARED_LIBS)
     add_definitions(-DUSE_IMPORT_EXPORT)
     # Copy relevant AWS SDK for C++ libraries into the current binary directory for running and debugging.
-    list(APPEND SERVICE_LIST dynamodb)
-    #For IDE's like Xcode and Visual Studio this line will be ignored because Release/Debug 
-    #  is switched internally, but this is necessary for non-IDE builds.
-    set(CMAKE_BUILD_TYPE Debug) #TODO: Set to your build type
-	
-    #TODO:Choose appropriate one of the following two lines, you want to copy to the same folder where your executables are.
-    AWSSDK_CPY_DYN_LIBS(SERVICE_LIST "" ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_BUILD_TYPE})  #Choose this line if your executables are in /build/Debug
-    #AWSSDK_CPY_DYN_LIBS(SERVICE_LIST "" ${CMAKE_CURRENT_BINARY_DIR})  #Choose this line for Visual Studio and possibly other IDEs
-    
-    message(STATUS ">>CMAKE_CURRENT_BINARY_DIR: ${CMAKE_CURRENT_BINARY_DIR}")
-    message(STATUS ">>CMAKE_BUILD_TYPE: ${CMAKE_BUILD_TYPE}")
+
+    set(CMAKE_BUILD_TYPE Debug) #Explicitly setting CMAKE_BUILD_TYPE is necessary in windows to copy dlls.
+
+    list(APPEND SERVICE_LIST ${SERVICE_COMPONENTS})
+
+     AWSSDK_CPY_DYN_LIBS(SERVICE_LIST "" ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_BUILD_TYPE})  #Choose this line if your executables are in /build/Debug
 endif()
 
-# Add the code example-specific header files.
-file(GLOB AWSDOC_S3_HEADERS
-    "include/awsdoc/s3/*.h"
-)
 
-# Add the code example-specific source files.
-file(GLOB AWSDOC_S3_SOURCE
-    "*.cpp"
-)
+# AWSDOC_${SERVICE_COMPONENTS}_SOURCE can be defined in the command line to limit the files in a build, for example to one action.
+if(NOT DEFINED AWSDOC_SOURCE)
+    file(GLOB AWSDOC_SOURCE
+        "*.cpp"
+    )
+endif()
 
 # Check whether the target system is Windows, including Win64.
 if(WIN32)
+    list(FILTER AWSDOC_SOURCE EXCLUDE REGEX "/list_buckets_disabling_dns_cache.cpp$") # Not supported in windows, see file for details
+
     # Check whether the compiler is some version of Microsoft Visual C++, or another compiler simulating C++.
     if(MSVC)
-        source_group("Header Files\\awsdoc\\s3" FILES ${AWSDOC_S3_HEADERS})
-        source_group("Source Files" FILES ${AWSDOC_S3_SOURCE})
+         source_group("Source Files" FILES ${AWSDOC_SOURCE})
     endif(MSVC)
 endif()
 
-foreach(file ${AWSDOC_S3_SOURCE})
+foreach(file ${AWSDOC_SOURCE})
     get_filename_component(EXAMPLE ${file} NAME_WE)
 
     # Build the code example executables.
     set(EXAMPLE_EXE run_${EXAMPLE})
 
-    add_executable(${EXAMPLE_EXE} ${AWSDOC_S3_HEADERS} ${file})
-
-    if(MSVC AND BUILD_SHARED_LIBS)
-        target_compile_definitions(${EXAMPLE_EXE} PUBLIC "USE_IMPORT_EXPORT")
-        target_compile_definitions(${EXAMPLE_EXE} PRIVATE "AWSDOC_S3_EXPORTS")
-    endif()
+    add_executable(${EXAMPLE_EXE}
+            ${file}
+            utilities/dynamodb_utilities.cpp
+            )
 
     target_include_directories(${EXAMPLE_EXE} PUBLIC 
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-        $<INSTALL_INTERFACE:include>)
+        $<INSTALL_INTERFACE:include>
+            ${ADDITIONAL_INCLUDE_DIR}
+            ${AWSSDK_INCLUDE_DIR}/aws
+            )
     target_link_libraries(${EXAMPLE_EXE} ${AWSSDK_LINK_LIBRARIES} 
-        ${AWSSDK_PLATFORM_DEPS})
+        ${AWSSDK_PLATFORM_DEPS}
+            ${ADDITIONAL_LIBS})
+    target_compile_definitions(${EXAMPLE_EXE}
+            PRIVATE
+            ${ADDITIONAL_DEFINES})
 
-    if(BUILD_TESTING)
-        # Enable testing for this directory and below.
-        enable_testing()
+ endforeach()
 
-        # Build the code example libraries.
-        set(EXAMPLE_LIB ${EXAMPLE}) 
+if(BUILD_TESTS)
+    add_subdirectory(gtests)
+endif()
 
-        add_library(${EXAMPLE_LIB} ${AWSDOC_S3_HEADERS} ${file} )
-
-        if(MSVC AND BUILD_SHARED_LIBS)
-            target_compile_definitions(${EXAMPLE_LIB} PUBLIC "USE_IMPORT_EXPORT")
-            target_compile_definitions(${EXAMPLE_LIB} PRIVATE "AWSDOC_S3_EXPORTS")
-        endif()
-
-        target_include_directories(${EXAMPLE_LIB} PUBLIC
-            $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-            $<INSTALL_INTERFACE:include>)
-        target_link_libraries(${EXAMPLE_LIB} ${AWSSDK_LINK_LIBRARIES} 
-            ${AWSSDK_PLATFORM_DEPS})
-
-        # Build the code example unit tests.
-        set(EXAMPLE_TEST test_${EXAMPLE})
-        set(EXAMPLE_TEST_FILE ${CMAKE_CURRENT_SOURCE_DIR}/tests/test_${EXAMPLE}.cpp)
-
-        if(EXISTS ${EXAMPLE_TEST_FILE})
-            add_executable(${EXAMPLE_TEST} ${EXAMPLE_TEST_FILE} )
-
-            target_include_directories(${EXAMPLE_TEST} PUBLIC
-                $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-                $<INSTALL_INTERFACE:include>)
-            target_link_libraries(${EXAMPLE_TEST} ${EXAMPLE_LIB} )
-            add_test(${EXAMPLE_TEST} ${EXAMPLE_TEST})
-        endif()
-
-    endif()
-endforeach()
